Remove WildCropFeatureMixin, move existing features to SpookyFeatures, move SpookyBiomeFeatures to the feature package.
