+++ src/main/java/com/fabriccommunity/spookytime/world/biome/SpookyBaseBiome.java
@@ -1,7 +1,8 @@
 package com.fabriccommunity.spookytime.world.biome;
 
+import com.fabriccommunity.spookytime.registry.SpookyFeatures;
+import com.fabriccommunity.spookytime.world.feature.SpookyBiomeFeatures;
 import com.google.common.collect.Lists;
-
 import net.minecraft.world.biome.Biome;
 import net.minecraft.world.biome.DefaultBiomeFeatures;
 import net.minecraft.world.gen.surfacebuilder.ConfiguredSurfaceBuilder;
@@ -12,7 +13,7 @@ import java.util.List;
 public abstract class SpookyBaseBiome extends Biome {
 	public static final List<Biome> BIOMES = Lists.newArrayList();
 	
-	protected static final ConfiguredSurfaceBuilder<?> SURFACE_BUILDER = new ConfiguredSurfaceBuilder<>(SurfaceBuilder.DEFAULT, SpookyBiomeFeatures.SPOOKY_FOREST);
+	protected static final ConfiguredSurfaceBuilder<?> SURFACE_BUILDER = new ConfiguredSurfaceBuilder<>(SurfaceBuilder.DEFAULT, SpookyFeatures.SPOOKY_FOREST);
 	
 	protected SpookyBaseBiome(Settings settings) {
 		super(settings.parent(null));
