+++ src/main/java/com/fabriccommunity/spookytime/world/biome/SpookySeaBiome.java
@@ -2,6 +2,7 @@ package com.fabriccommunity.spookytime.world.biome;
 
 import com.fabriccommunity.spookytime.registry.SpookyBlocks;
 
+import com.fabriccommunity.spookytime.world.feature.SpookyBiomeFeatures;
 import net.minecraft.entity.EntityCategory;
 import net.minecraft.entity.EntityType;
 import net.minecraft.util.math.BlockPos;
